IceCream in new class file
